+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Notifications on reactions received (configurable).
 - Setting to turn off sending a message when the enter-key is pressed. 
+- Ability to submit captcha requests.
 
 ### Changed